remove chapters logic from multisrc class since all three have different logic
