+++ lib-multisrc/gmanga/src/eu/kanade/tachiyomi/multisrc/gmanga/Dto.kt
@@ -1,12 +1,8 @@
 package eu.kanade.tachiyomi.multisrc.gmanga
 
-import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.float
-import java.text.SimpleDateFormat
 
 @Serializable
 class EncryptedResponse(val data: String)
@@ -127,34 +123,6 @@ class TypeDto(
     val name: String,
     val title: String,
 )
-
-@Serializable
-class ChapterListDto(
-    val releases: List<ChapterRelease>,
-)
-
-@Serializable
-class ChapterRelease(
-    private val id: Int,
-    private val chapter: JsonPrimitive,
-    private val title: String,
-    @SerialName("team_name") private val teamName: String,
-    @SerialName("created_at") private val createdAt: String,
-) {
-    fun toSChapter(dateFormat: SimpleDateFormat) = SChapter.create().apply {
-        url = "/r/$id"
-        chapter_number = chapter.float
-        date_upload = try {
-            dateFormat.parse(createdAt)!!.time
-        } catch (_: Exception) {
-            0L
-        }
-        scanlator = teamName
-
-        val chapterName = title.let { if (it.trim() != "") " - $it" else "" }
-        name = "${chapter_number.let { if (it % 1 > 0) it else it.toInt() }}$chapterName"
-    }
-}
 
 @Serializable
 class ReaderDto(
