+++ src/ar/dilar/src/eu/kanade/tachiyomi/extension/ar/dilar/Dilar.kt
@@ -1,7 +1,10 @@
 package eu.kanade.tachiyomi.extension.ar.dilar
 
 import eu.kanade.tachiyomi.multisrc.gmanga.Gmanga
+import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.source.model.SChapter
+import eu.kanade.tachiyomi.source.model.SManga
+import okhttp3.Request
 import okhttp3.Response
 
 class Dilar : Gmanga(
@@ -9,12 +12,15 @@ class Dilar : Gmanga(
     "https://dilar.tube",
     "ar",
 ) {
-    override fun chapterListParse(response: Response): List<SChapter> {
+    override fun chaptersRequest(manga: SManga): Request {
+        val mangaId = manga.url.substringAfterLast("/")
+        return GET("$baseUrl/api/mangas/$mangaId/releases", headers)
+    }
+
+    override fun chaptersParse(response: Response): List<SChapter> {
         val releases = response.parseAs<ChapterListDto>().releases
             .filterNot { it.isMonetized }
 
-        return releases.map {
-            it.toSChapter(dateFormat)
-        }.sortChapters()
+        return releases.map { it.toSChapter() }
     }
 }
