+++ src/ar/mangatales/src/eu/kanade/tachiyomi/extension/ar/mangatales/Dto.kt
@@ -1,13 +1,42 @@
 package eu.kanade.tachiyomi.extension.ar.mangatales
 
-import eu.kanade.tachiyomi.multisrc.gmanga.ChapterRelease
+import eu.kanade.tachiyomi.source.model.SChapter
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.JsonPrimitive
+import kotlinx.serialization.json.float
+import java.text.SimpleDateFormat
+import java.util.Locale
 
 @Serializable
 class ChapterListDto(
     val mangaReleases: List<ChapterRelease>,
 )
+
+@Serializable
+class ChapterRelease(
+    private val id: Int,
+    private val chapter: JsonPrimitive,
+    private val title: String,
+    @SerialName("team_name") private val teamName: String,
+    @SerialName("created_at") private val createdAt: String,
+) {
+    fun toSChapter() = SChapter.create().apply {
+        url = "/r/$id"
+        chapter_number = chapter.float
+        date_upload = try {
+            dateFormat.parse(createdAt)!!.time
+        } catch (_: Exception) {
+            0L
+        }
+        scanlator = teamName
+
+        val chapterName = title.let { if (it.trim() != "") " - $it" else "" }
+        name = "${chapter_number.let { if (it % 1 > 0) it else it.toInt() }}$chapterName"
+    }
+}
+
+private val dateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", Locale.ENGLISH)
 
 @Serializable
 class ReaderDto(
